refactor: move logic to hook
